requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
